add_footnote {kableExtra}	R Documentation

Add footnote

Description

Add footnote to your favorite kable output.

Usage

add_footnote(input, label = NULL, notation = "alphabet",
  threeparttable = FALSE, escape = TRUE)

Arguments

input
The direct output of your kable function or your last kableExtra function.

label
A vector of footnotes you want to add. You don't need to add notations in your notes.

notation
You can select the format of your footnote notation from number, alphabet, symbol and none.

threeparttable
Boolean value indicating if a threeparttable scheme should be used.

escape
Logical value controlling if the label needs to be escaped. Default is TRUE.

Examples

x <- knitr::kable(head(mtcars), "html")
add_footnote(x, c("footnote 1", "footnote 2"), notation = "symbol")

[Package kableExtra version 1.1.0 Index]