as_image {kableExtra}	R Documentation

Render the table as an format-independent image and use it in R Markdown

Description

This function generates a temporary png file using save_kable and then try to put it in an R Markdown document using knitr::include_graphics.

Usage

as_image(x, width = NULL, height = NULL, file = NULL, ...)

Arguments

x
kable input. Either HTML or LaTeX

width
Image width in inches. (1 inch = 2.54 cm)

height
Image height in inches. (1 inch = 2.54 cm)

file
By default, as_image saves to an temp file, which works for normal R Markdown. However if you are using things like xaringan, which can't be a standalone html, you can specify this file be the path you need, e.g. "img/something.png"

...
Additional arguments passed to save_kable.

Examples

## Not run: 
library(kableExtra)

kable(mtcars, "latex", booktabs = T) %>%
kable_styling(latex_options = c("striped", "scale_down")) %>%
row_spec(1, color = "red") %>%
as_image()

## End(Not run)

[Package kableExtra version 1.4.0 Index]