auto_index {kableExtra}	R Documentation

Automatically figuring out the group_row index

Description

This helper function allows users to build the group_row index more quickly and use group_rows in a way that is similar with collapse_rows.

Usage

auto_index(x)

Arguments

x
The index column. A vector. For example 'c("a", "a", "b", "b", "b")“

[Package kableExtra version 1.1.0 Index]