footnote_marker_number {kableExtra}	R Documentation

Footnote marker

Description

Put footnote mark in superscription in table. Unless you are using it in the caption of kable, you will need to put escape = F in kable (similar with cell_spec). Again, similar with cell_spec, the format option here can read default value from global option knitr.table.format.

Usage

footnote_marker_number(x, format, double_escape = FALSE)

footnote_marker_alphabet(x, format, double_escape = FALSE)

footnote_marker_symbol(x, format, double_escape = FALSE)

Arguments

x
a number. For example, for footnote_marker_alphabet(2) will return "b" in HTML.

format
Either html or latex. All functions here can read default value from global option knitr.table.format.

double_escape
T/F if output is in LaTeX, whether it should be double escaped. If you are using footnote_marker in group_rows`` labeling row or add_header_above, you need to set this to be TRUE'.

Examples

## Not run: 
dt <- mtcars[1:5, 1:5]
colnames(dt)[1] <- paste0("mpg",
                          footnote_marker_alphabet(2))
rownames(dt)[2] <- paste0(rownames(dt)[2],
                          footnote_marker_alphabet(1))
dt[1,2] <- paste0(dt[1,2], footnote_marker_alphabet(3))

kbl(dt, escape = FALSE) |>
  footnote(alphabet = c("Note a", "Note b", "Note c"))

## End(Not run)

[Package kableExtra version 1.4.0 Index]