header_separate {kableExtra}	R Documentation

Separate table headers and add additional header rows based on grouping

Description

When you create a summary table for either model or basic summary stats in R, you usually end up having column names in the form of "a_mean", "a_sd", "b_mean" and "b_sd". This function streamlines the process of renaming these column names and adding extra header rows using add_header_above.

Usage

header_separate(kable_input, sep = "[^[:alnum:]]+", ...)

Arguments

kable_input
Output of knitr::kable() with format specified

sep
A regular expression separator between groups. The default value is a regular expression that matches any sequence of non-alphanumeric values.

...
Additional parameters passed to do.call.

[Package kableExtra version 1.4.0 Index]