kable_as_image {kableExtra}	R Documentation

Deprecated

Description

deprecated

Usage

kable_as_image(kable_input, filename = NULL, file_format = "png",
  latex_header_includes = NULL, keep_pdf = FALSE, density = 300,
  keep_tex = FALSE)

Arguments

kable_input
Raw LaTeX code to generate a table. It doesn't have to came from kable or kableExtra.

filename
Character String. If specified, the image will be saved under the specified (path &) name. You don't need to put file format like ".png" here.

file_format
Character String to specify image format, such as png, jpeg, gif, tiff, etc. Default is png.

latex_header_includes
A character vector of extra LaTeX header stuff. Each element is a row. You can have things like c("\\usepackage{threeparttable}", "\\usepackage{icons}") You could probably add your language package here if you use non-English text in your table, such as \\usepackage[magyar]{babel}.

keep_pdf
A T/F option to control if the mid-way standalone pdf should be kept. Default is FALSE.

density
Resolution to read the PDF file. Default value is 300, which should be sufficient in most cases.

keep_tex
A T/F option to control if the latex file that is initially created should be kept. Default is FALSE.

[Package kableExtra version 1.1.0 Index]