kable_classic {kableExtra}	R Documentation

Alternative HTML themes

Description

kableExtra uses the built-in bootstrap themes by default in kable_styling(). Alternatively, you can use a customized table themes for your table. This lightable table style sheet comes with three formats, namely lightable-minimal, lightable-classic, lightable-material and lightable-material-dark with hover and striped options.

Usage

kable_classic(
  kable_input,
  lightable_options = "basic",
  html_font = "\"Arial Narrow\", \"Source Sans Pro\", sans-serif",
  ...
)

kable_classic_2(
  kable_input,
  lightable_options = "basic",
  html_font = "\"Arial Narrow\", \"Source Sans Pro\", sans-serif",
  ...
)

kable_minimal(
  kable_input,
  lightable_options = "basic",
  html_font = "\"Trebuchet MS\", verdana, sans-serif",
  ...
)

kable_material(
  kable_input,
  lightable_options = "basic",
  html_font = "\"Source Sans Pro\", helvetica, sans-serif",
  ...
)

kable_material_dark(
  kable_input,
  lightable_options = "basic",
  html_font = "\"Source Sans Pro\", helvetica, sans-serif",
  ...
)

kable_paper(
  kable_input,
  lightable_options = "basic",
  html_font = "\"Arial Narrow\", arial, helvetica, sans-serif",
  ...
)

Arguments

kable_input
A HTML kable object.

lightable_options
Options to customize lightable. Similar with bootstrap_options in kable_styling. Choices include basic, striped and hover.

html_font
A string for HTML css font. For example, html_font = '"Arial Narrow", arial, helvetica, sans-serif'.

...
Everything else you need to specify in kable_styling.

[Package kableExtra version 1.4.0 Index]