landscape {kableExtra}	R Documentation

Print the table on an isolated landscape page in PDF

Description

This function will put the table on an single landscape page. It's useful for wide tables that cann't be printed on a portrait page.

Usage

landscape(kable_input, margin = NULL)

Arguments

kable_input
Output of knitr::kable() with format specified

margin
Customizable page margin for special needs. Values can be "1cm", "1in" or similar.

Examples

landscape(knitr::kable(head(mtcars), "latex"))

[Package kableExtra version 1.1.0 Index]