make_inline_plot {kableExtra}	R Documentation

Combine file (or svg text) and parameters into a kableExtraInlinePlots object

Description

Combine file (or svg text) and parameters into a kableExtraInlinePlots object

Usage

make_inline_plot(filename, file_ext, dev, width, height, res, del = TRUE)

Arguments

filename
Passed through to the graphics device.

file_ext
Character, something like "png".

dev
Character (e.g., "svg", "pdf") or function (e.g.,

width, height
Plot dimensions in pixels.

res
The resolution of the plot; default is 300.

del
If the file is svg-like, then the default action is to read the file into an embedded SVG object; once done, the file is no longer used. The default action is to delete this file early, set this to 'FALSE' to keep the file.

Value

list object, with class kableExtraInlinePlots

[Package kableExtra version 1.4.0 Index]