spec_image {kableExtra}	R Documentation

Setup image path, size, etc

Description

Users can directly provide image file path to column spec. However, if you need to specify the size of the image, you will need this function.

Usage

spec_image(path, width, height, res = 300, svg_text = NULL)

Arguments

path
file path(s)

width
image width in pixel

height
image height in pixel

res
image resolution.

svg_text
If you have the raw text for SVG. Put them here

[Package kableExtra version 1.4.0 Index]