xtable2kable {kableExtra}	R Documentation

Convert xtable to a kable object

Description

This function allow users to turn an xtable object into a kable so they can use most of kableExtra's functions with their xtable code without making too many changes. Note that although I tested many cases and it seems to work, this function may not be functional in some other cases. I'm not a regular xtable user and can only provide very limited support for this function.

You should use this table in the same way as print.xtable. All the options you provided to this function will be sent to print.xtable. Instead of printing out the result, this function will return the LaTeX or HTML as text and a kable object.

Usage

xtable2kable(x, ...)

Arguments

x
an xtable object

...
options for print.xtable

Examples

## Not run: 
library(xtable)
xtable(mtcars) %>%
  xtable2kable(booktabs = TRUE) %>%
  kable_styling(latex_options = "striped")

## End(Not run)

[Package kableExtra version 1.1.0 Index]