save_kable {kableExtra}	R Documentation

Save kable to files

Description

Save kable to files

Usage

save_kable(x, file, bs_theme = "simplex", self_contained = TRUE,
  extra_dependencies = NULL, ..., latex_header_includes = NULL,
  keep_tex = FALSE)

Arguments

x
A piece of HTML code for tables, usually generated by kable and kableExtra

file
save to files. If the input table is in HTML and the output file ends with .png, .pdf and .jpeg, webshot will be used to do the conversion.

bs_theme
Which Bootstrap theme to use

self_contained
Will the files be self-contained?

extra_dependencies
Additional HTML dependencies. For example, list(

...
Additional variables being passed to webshot::webshot. This is for HTML only.

latex_header_includes
A character vector of extra LaTeX header stuff. Each element is a row. You can have things like c("\\usepackage{threeparttable}", "\\usepackage{icons}") You could probably add your language package here if you use non-English text in your table, such as \\usepackage[magyar]{babel}.

keep_tex
A T/F option to control if the latex file that is initially created should be kept. Default is FALSE.

[Package kableExtra version 1.1.0 Index]